potali-5-1 add cancan

+++ app/helpers/questions_helper.rb
@@ -1,7 +1,8 @@
 module QuestionsHelper
-  	def question_actions(user, question)
-    	html = []
-		if question.user_id == user.id
+  def question_actions(user, question)
+    if user.present?
+      html = []
+  		if question.user_id == user.id
         	html << link_to('', edit_question_path(question), :title => "Edit", :class=>'icon-edit')
         	html << link_to('', question, method: :delete, :title => "Delete", :class=>'icon-remove', data: { confirm: 'Are you sure?',:remote=>true })
     	    if Question.already_published(question)
@@ -16,6 +17,7 @@ module QuestionsHelper
           		html << link_to("", flag_question_path(question), :remote => true, :class => "icon-flag" ,:id=>"flag_#{question.id}")
         	end
     	end
-  		raw html.join(" ")
-  	end
+    	raw html.join(" ")
+    end
+  end
 end
